+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 ### Fixed
+- Fixed the disabled state CSS for the `cancel` All button, removed the `cancel all` button from the ProgressBar subheader, and removed focus from cancelled items in the ProgressBar.
 - Fixed tab indicator not updating when tabValue changes in `Tabs` component
 
 ### Changed